[ADMIN] Re-organize organization components
