+++ admin/app/templates/authenticated/organizations/get.hbs
@@ -8,7 +8,7 @@
 
 <main class="page-body" id="organizations-get-page">
 
-  <OrganizationInformationSection
+  <Organizations::InformationSection
     @organization={{@model}}
     @onLogoUpdated={{this.updateOrganizationInformation}}
     @onSubmit={{this.updateOrganizationInformation}}
