+++ admin/app/templates/authenticated/organizations/get/all-tags.hbs
@@ -1,2 +1,2 @@
 {{page-title "Orga " @model.organization.id " | Tags"}}
-<OrganizationAllTags @model={{@model}} />
+<Organizations::AllTags @model={{@model}} />
