+++ admin/app/templates/authenticated/organizations/get/campaigns.hbs
@@ -1,2 +1,2 @@
 {{page-title "Orga " @model.organizationId " | Campagnes"}}
-<OrganizationCampaignsSection @campaigns={{@model.campaigns}} />
+<Organizations::CampaignsSection @campaigns={{@model.campaigns}} />
