+++ admin/app/templates/authenticated/organizations/get/target-profiles.hbs
@@ -1,2 +1,2 @@
 {{page-title "Orga " @model.id " | Profils cibles"}}
-<OrganizationTargetProfilesSection @organization={{@model}} />
+<Organizations::TargetProfilesSection @organization={{@model}} />
