+++ admin/app/templates/authenticated/organizations/new.hbs
@@ -7,7 +7,7 @@
 
 <main class="page-body">
   <section class="page-section">
-    <OrganizationForm
+    <Organizations::Form
       @organization={{@model}}
       @onSubmit={{this.addOrganization}}
       @onCancel={{this.goBackToOrganizationList}}
