+++ admin/tests/unit/components/organizations/information-section_test.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
-import createGlimmerComponent from '../../helpers/create-glimmer-component';
+import createGlimmerComponent from '../../../helpers/create-glimmer-component';
 import ENV from 'pix-admin/config/environment';
 
 module('Unit | Component | organization-information-section', function (hooks) {
@@ -9,7 +9,7 @@ module('Unit | Component | organization-information-section', function (hooks) {
   let component;
 
   hooks.beforeEach(function () {
-    component = createGlimmerComponent('component:organization-information-section');
+    component = createGlimmerComponent('component:organizations/information-section');
   });
 
   test('it should generate link based on environment and object', async function (assert) {
